Updated: revisi inventory, Production, Data  (populasi)

+++ Models/Inventory/Barang.cs
@@ -4,7 +4,7 @@ namespace CakeProject.Models.Inventory
     {
         public int IDBarang { get; set; }
         public string NamaBarang { get; set; }
-        public double HargaBeli { get; set; }
+        public int HargaBeli { get; set; }
         public double JumlahStok { get; set; }
         public double MinimalStok { get; set; }
         public string Satuan { get; set; }
